+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-# v1.4.2
+## v1.5.0
+
+* (#206) Support multiple emits for the same timestep and remove
+  `Engine.complete()`. Instead, callers of `Engine.run_for()` must pass
+  `force_complete=True` at the end of their caller-managed simulation
+  loops.
+
+## v1.4.2
 
 * (#203) Make `Store.soures` more comprehensive, in particular by
   including dividers and flows.